parse-options: use and require int pointer for OPT_CMDMODE
Some uses of OPT_CMDMODE provide a pointer to an enum.  It is
dereferenced as an int pointer in parse-options.c::get_value().  The two
types are incompatible, though -- the storage size of an enum can vary
between platforms.  C23 would allow us to specify the underlying type of
the diffenrent enums, making them compatible, but with C99 the easiest
safe option is to actually use int as the value type.

Convert the offending OPT_CMDMODE users to point to a new int value and
set the enum value after parsing.  Switch to the typed value_int pointer
in the macro's definition to enforce that type for future users.

Signed-off-by: René Scharfe <l.s.r@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>